Put server data files under ``~ereuse/data``, bind mount it to ``/srv/ereuse-data``
Why export the bind mount and not the original directory?  Because it
enables us to cache files in other places (e.g. a ``tmpfs``) and then
bind mount the cached file into ``/srv/ereuse-data``, so clients access
the cached file, but the original one is still available under
``~ereuse/data``.