Add script to refresh exported ISOs.
The script locates ISOs in the ``images`` subdirectory of the shared
directory and automatically mounts and exports the mountpoint via NFS.
If a ``.syslinux`` template accompanies the ISO, its variables are
resolved and the result is added to PXE configuration so that the entry
appears in the PXELINUX menu.

Still need to test it and include it in the boot process.